refractor(db-schema): change points of activities to decimal

+++ schemas/db-schema.sql
@@ -43,7 +43,7 @@ CREATE TABLE activities(
     activities_id int NOT NULL PRIMARY KEY AUTO_INCREMENT,
     name varchar(30) NOT NULL,
     description varchar(100) NULL,
-    points int NOT NULL,
+    points decimal NOT NULL,
     UNIQUE KEY (activities_name)
 );
